Refactor and update various components
- Changed default value of `opt->search_depth` in `main.cpp` from `ULLONG_MAX` to `INT_MAX`.
- Enhanced `scalar.h` with new includes, inline functions for field initialization and setting, and a template function for binary search.
- Replaced `memset` with `s_memset` in `snmod_mat.cpp`.
- Removed several includes from `sparse_vec.h`.
- Removed `util.cpp` and `util.h`.
- Introduced `base.h` to consolidate functionalities from `util.h`, including memory management, field structure definitions, and utility functions for strings and time measurement.

+++ main.cpp
@@ -187,7 +187,7 @@ int main(int argc, char** argv) {
 	opt->search_min = program.get<int>("--search_min");
 	opt->pivot_dir = (program.get<std::string>("--pivot_direction") == "row");
 	if (opt->search_depth == 0)
-		opt->search_depth = ULLONG_MAX;
+		opt->search_depth = INT_MAX;
 	if (opt->sort_step == 0)
 		opt->sort_step = (int)std::max((ulong)1000, n_col / 100);
